# contract/openapi.yaml
...
components:
  schemas:
...
    AdminTaskResponse:
      properties:
        attempts:
          format: "int32"
          type: "integer"
        createdAt:
          format: "date-time"
          type: "string"
        currentStep:
          format: "int32"
          type: "integer"
        hostname:
          type:
          - "string"
          - "null"
        jobrunrJobId:
          type:
          - "string"
          - "null"
        kind:
          $ref: "#/components/schemas/ProvisioningTaskKind"
        lastError:
          type:
          - "string"
          - "null"
        orgId:
          format: "uuid"
          type:
          - "string"
          - "null"
        orgName:
          type:
          - "string"
          - "null"
        status:
          $ref: "#/components/schemas/ProvisioningTaskStatus"
        stepLabel:
          type: "string"
        taskId:
          format: "uuid"
          type: "string"
        totalSteps:
          format: "int32"
          type: "integer"
        updatedAt:
          format: "date-time"
          type: "string"
        vmId:
          format: "uuid"
          type: "string"
        vmName:
          type:
          - "string"
          - "null"
        workspaceName:
          type:
          - "string"
          - "null"
      required:
      - "attempts"
      - "createdAt"
      - "currentStep"
      - "kind"
      - "status"
      - "stepLabel"
      - "taskId"
      - "totalSteps"
      - "updatedAt"
      - "vmId"
      type: "object"
...
    Attention:
      properties:
        expiredVmCount:
          format: "int64"
          type: "integer"
        failedTaskCount:
          format: "int64"
          type: "integer"
        needsAdminVmCount:
          format: "int64"
          type: "integer"
      required:
      - "expiredVmCount"
      - "failedTaskCount"
      - "needsAdminVmCount"
      type: "object"
...
    CapacityTrendPointResponse:
      properties:
        day:
          format: "date"
          type: "string"
        diskGb:
          format: "int64"
          type: "integer"
        memoryMb:
          format: "int64"
          type: "integer"
        vcpu:
          format: "int64"
          type: "integer"
        vmCount:
          format: "int64"
          type: "integer"
      required:
      - "day"
      - "diskGb"
      - "memoryMb"
      - "vcpu"
      - "vmCount"
      type: "object"
    CapacityTrendResponse:
      properties:
        capacityCpuThreads:
          format: "int64"
          type: "integer"
        capacityDiskGb:
          description: "ACTIVE 노드의 thin pool 용량 합(GB) — 오버프로비저닝 전제의 조언용 분모, 용량 미등록\
            \ 노드가 있으면 null"
          format: "int64"
          type:
          - "integer"
          - "null"
        capacityMemoryMb:
          format: "int64"
          type: "integer"
        from:
          format: "date"
          type: "string"
        points:
          description: "일 단위 스냅샷 — 해당 일 종료 시점에 살아 있던 VM의 할당 합"
          items:
            $ref: "#/components/schemas/CapacityTrendPointResponse"
          type: "array"
        to:
          format: "date"
          type: "string"
      required:
      - "capacityCpuThreads"
      - "capacityMemoryMb"
      - "from"
      - "points"
      - "to"
      type: "object"
...
    DisableMfaRequest:
      properties:
        code:
          pattern: "^[0-9]{6}$"
          type: "string"
        password:
          minLength: 1
          type: "string"
        recoveryCode:
          type: "string"
      required:
      - "password"
      type: "object"
    DisableUserRequest:
      properties:
        reason:
          maxLength: 500
          minLength: 1
          type: "string"
      required:
      - "reason"
      type: "object"
...
    FieldValidationError:
      description: "검증 실패 필드 한 건 (422 응답의 errors[] 원소)"
      properties:
        field:
          description: "실패한 요청 필드 이름"
          type: "string"
        message:
          description: "한국어 오류 메시지"
          type: "string"
      required:
      - "field"
      - "message"
      type: "object"
...
    IpPoolUsage:
      properties:
        allocatedCount:
          format: "int64"
          type: "integer"
        cidr:
          type: "string"
        freeCount:
          format: "int64"
          type: "integer"
        id:
          format: "uuid"
          type: "string"
        name:
          type: "string"
      required:
      - "allocatedCount"
      - "cidr"
      - "freeCount"
      - "id"
      - "name"
      type: "object"
    LiveCoverage:
      properties:
        memoryMeasuredNodeCount:
          description: "메모리 측정값이 있는 노드 수 — nodeCount보다 작으면 메모리 합계는 부분 측정"
          format: "int32"
          type: "integer"
        nodeCount:
          description: "nodesLive 행 수 (플랫폼 전체 노드 수)"
          format: "int32"
          type: "integer"
        storageMeasuredNodeCount:
          description: "스토리지 측정값이 있는 노드 수 — nodeCount보다 작으면 스토리지 합계는 부분 측정"
          format: "int32"
          type: "integer"
      required:
      - "memoryMeasuredNodeCount"
      - "nodeCount"
      - "storageMeasuredNodeCount"
      type: "object"
...
    MessageResponse:
      properties:
        message:
          type: "string"
      required:
      - "message"
      type: "object"
...
    NodeLiveResponse:
      properties:
        checkedAt:
          format: "date-time"
          type:
          - "string"
          - "null"
        cpu:
          format: "double"
          type:
          - "number"
          - "null"
        memTotalBytes:
          format: "int64"
          type:
          - "integer"
          - "null"
        memUsedBytes:
          format: "int64"
          type:
          - "integer"
          - "null"
        name:
          type: "string"
        nodeId:
          format: "uuid"
          type: "string"
        reachable:
          description: "true = 이 노드가 상태 조회에 응답함. false = 응답하지 않음이며 나머지 필드는 모두 null.\
            \ true여도 스토리지 조회는 별도 권한이라 storage* 필드는 null일 수 있음"
          type: "boolean"
        storageTotalBytes:
          format: "int64"
          type:
          - "integer"
          - "null"
        storageUsedBytes:
          format: "int64"
          type:
          - "integer"
          - "null"
      required:
      - "name"
      - "nodeId"
      - "reachable"
      type: "object"
...
    NodeRatio:
      properties:
        cpuOvercommitRatio:
          format: "double"
          type: "number"
        id:
          format: "uuid"
          type: "string"
        memoryAllocRatio:
          format: "double"
          type: "number"
        name:
          type: "string"
        status:
          $ref: "#/components/schemas/NodeStatus"
        warn:
          type: "boolean"
      required:
      - "cpuOvercommitRatio"
      - "id"
      - "memoryAllocRatio"
      - "name"
      - "status"
      - "warn"
      type: "object"
    NodeStatus:
      enum:
      - "ACTIVE"
      - "MAINTENANCE"
      - "OFFLINE"
      type: "string"
...
    OrgDashboardSummaryResponse:
      properties:
        attention:
          $ref: "#/components/schemas/Attention"
        expiringVmCount30d:
          format: "int64"
          type: "integer"
        pendingRequestCount:
          format: "int64"
          type: "integer"
        publishedServiceCount:
          format: "int64"
          type: "integer"
        recentDecisions14d:
          $ref: "#/components/schemas/RecentDecisions"
        resource:
          $ref: "#/components/schemas/Resource"
        topWorkspacesByVmCount:
          items:
            $ref: "#/components/schemas/TopWorkspace"
          type: "array"
        vmCountsByStatus:
          additionalProperties:
            format: "int64"
            type: "integer"
          type: "object"
      required:
      - "attention"
      - "expiringVmCount30d"
      - "pendingRequestCount"
      - "publishedServiceCount"
      - "recentDecisions14d"
      - "resource"
      - "topWorkspacesByVmCount"
      - "vmCountsByStatus"
      type: "object"
...
    PageResponseAdminTaskResponse:
      properties:
        content:
          items:
            $ref: "#/components/schemas/AdminTaskResponse"
          type: "array"
        page:
          format: "int32"
          type: "integer"
        size:
          format: "int32"
          type: "integer"
        totalElements:
          format: "int64"
          type: "integer"
        totalPages:
          format: "int32"
          type: "integer"
      required:
      - "content"
      - "page"
      - "size"
      - "totalElements"
      - "totalPages"
      type: "object"
...
    PageResponseUserAdminViewResponse:
      properties:
        content:
          items:
            $ref: "#/components/schemas/UserAdminViewResponse"
          type: "array"
        page:
          format: "int32"
          type: "integer"
        size:
          format: "int32"
          type: "integer"
        totalElements:
          format: "int64"
          type: "integer"
        totalPages:
          format: "int32"
          type: "integer"
      required:
      - "content"
      - "page"
      - "size"
      - "totalElements"
      - "totalPages"
      type: "object"
...
    Problem:
      description: "RFC 9457 오류 응답 + Pickle 확장(code, errors). 모든 비정상 응답은 이 형태로 반환됩\
        니다."
      properties:
        code:
          description: "기계 판독용 안정 오류 코드 (클라이언트 분기 기준)"
          type: "string"
        detail:
          description: "이 발생 건에 대한 설명"
          type: "string"
        errors:
          description: "검증 실패 상세 (422에서만 존재)"
          items:
            $ref: "#/components/schemas/FieldValidationError"
          type: "array"
        instance:
          description: "발생 경로"
          type: "string"
        status:
          description: "HTTP 상태 코드"
          format: "int32"
          type: "integer"
        title:
          description: "사람이 읽는 짧은 제목"
          type: "string"
        type:
          description: "문제 유형 URI (기본 about:blank)"
          type: "string"
      required:
      - "title"
      - "status"
      - "code"
      type: "object"
    ProvisioningTaskKind:
      enum:
      - "PROVISION"
      - "DELETE"
      - "REINSTALL"
      type: "string"
...
    ProvisioningTaskStatus:
      enum:
      - "PENDING"
      - "RUNNING"
      - "DONE"
      - "FAILED"
      - "RETRYING"
      - "NEEDS_ADMIN"
      type: "string"
...
    RecentDecisions:
      properties:
        approvedCount:
          format: "int64"
          type: "integer"
        rejectedCount:
          format: "int64"
          type: "integer"
      required:
      - "approvedCount"
      - "rejectedCount"
      type: "object"
...
    Resource:
      properties:
        allocatedDiskGb:
          format: "int64"
          type: "integer"
        allocatedMemoryMb:
          format: "int64"
          type: "integer"
        allocatedVcpu:
          format: "int64"
          type: "integer"
        capacityDiskGb:
          description: "ACTIVE 노드의 thin pool 용량 합(GB) — 용량 미등록 노드가 있으면 null"
          format: "int64"
          type:
          - "integer"
          - "null"
        capacityMemoryMb:
          format: "int64"
          type:
          - "integer"
          - "null"
        capacityVcpu:
          format: "int64"
          type:
          - "integer"
          - "null"
        guidance:
          type: "string"
      required:
      - "allocatedDiskGb"
      - "allocatedMemoryMb"
      - "allocatedVcpu"
      - "guidance"
      type: "object"
...
    SystemDashboardSummaryResponse:
      properties:
        certExpiring30dCount:
          format: "int64"
          type: "integer"
        ipPools:
          items:
            $ref: "#/components/schemas/IpPoolUsage"
          type: "array"
        liveCoverage:
          $ref: "#/components/schemas/LiveCoverage"
        nodes:
          items:
            $ref: "#/components/schemas/NodeRatio"
          type: "array"
        nodesLive:
          items:
            $ref: "#/components/schemas/NodeLiveResponse"
          type: "array"
        notificationFailureCount:
          format: "int64"
          type: "integer"
        openDriftFindingCount:
          format: "int64"
          type: "integer"
        sshPasswordEnabledVmCount:
          format: "int64"
          type: "integer"
        tasks:
          $ref: "#/components/schemas/Tasks"
        vmCountsByStatus:
          additionalProperties:
            format: "int64"
            type: "integer"
          type: "object"
      required:
      - "certExpiring30dCount"
      - "ipPools"
      - "liveCoverage"
      - "nodes"
      - "nodesLive"
      - "notificationFailureCount"
      - "openDriftFindingCount"
      - "sshPasswordEnabledVmCount"
      - "tasks"
      - "vmCountsByStatus"
      type: "object"
...
    Tasks:
      properties:
        failed24hCount:
          format: "int64"
          type: "integer"
        needsAdminCount:
          format: "int64"
          type: "integer"
        retryingCount:
          format: "int64"
          type: "integer"
        runningCount:
          format: "int64"
          type: "integer"
      required:
      - "failed24hCount"
      - "needsAdminCount"
      - "retryingCount"
      - "runningCount"
      type: "object"
...
    TopWorkspace:
      properties:
        name:
          type: "string"
        vmCount:
          format: "int64"
          type: "integer"
        workspaceId:
          format: "uuid"
          type: "string"
      required:
      - "name"
      - "vmCount"
      - "workspaceId"
      type: "object"
...
    UpdateUserAdminRequest:
      properties:
        orgId:
          format: "uuid"
          type:
          - "string"
          - "null"
        role:
          $ref: "#/components/schemas/UserRole"
      type: "object"
...
    UserAdminDetailResponse:
      properties:
        activeVmCount:
          format: "int32"
          type: "integer"
        createdAt:
          format: "date-time"
          type: "string"
        disabledAt:
          format: "date-time"
          type:
          - "string"
          - "null"
        disabledReason:
          type:
          - "string"
          - "null"
        email:
          type: "string"
        id:
          format: "uuid"
          type: "string"
        memberships:
          items:
            $ref: "#/components/schemas/Membership"
          type: "array"
        mfaEnabled:
          type: "boolean"
        name:
          type: "string"
        orgId:
          format: "uuid"
          type:
          - "string"
          - "null"
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
        statusChanges:
          items:
            $ref: "#/components/schemas/UserStatusChangeResponse"
          type: "array"
        withdrawnAt:
          format: "date-time"
          type:
          - "string"
          - "null"
      required:
      - "activeVmCount"
      - "createdAt"
      - "email"
      - "id"
      - "memberships"
      - "mfaEnabled"
      - "name"
      - "role"
      - "status"
      - "statusChanges"
      type: "object"
...
    UserRole:
      enum:
      - "USER"
      - "ORG_MANAGER"
      - "ORG_ADMIN"
      - "SYS_MANAGER"
      - "SYS_ADMIN"
      type: "string"
    UserStatus:
      enum:
      - "PENDING_VERIFICATION"
      - "ACTIVE"
      - "DISABLED"
      - "WITHDRAWN"
      type: "string"
...
    UserSummaryResponse:
      properties:
        email:
          type: "string"
        id:
          format: "uuid"
          type: "string"
        name:
          type: "string"
        role:
          $ref: "#/components/schemas/UserRole"
      required:
      - "email"
      - "id"
      - "name"
      - "role"
      type: "object"
...
  securitySchemes:
    bearerAuth:
      bearerFormat: "JWT"
      description: "액세스 토큰 (JWT HS256, 15분 만료). 클레임: sub(계정 공개 식별자 UUID), role, token_version.\
        \ 비밀번호 변경·계정 비활성화 시 token_version이 올라가 기존 토큰이 즉시 무효화됩니다. 리프레시 토큰은 보안 스킴이 아\
        니라 __Host-pickle_refresh httpOnly 쿠키로만 오갑니다."
      scheme: "bearer"
      type: "http"
info:
  description: "부산대학교 클라우드 플랫폼 Pickle의 REST API. 인증은 JWT Bearer, 오류 응답은 RFC 9457 problem+json(Problem\
    \ 스키마)을 따릅니다."
  title: "Pickle API"
  version: "0.38.0"
openapi: "3.1.0"
paths:
...
  /admin/capacity-trend:
    get:
      operationId: "getAdminCapacityTrend"
      parameters:
      - in: "query"
        name: "days"
        required: false
        schema:
          default: 90
          format: "int32"
          maximum: 365
          minimum: 7
          type: "integer"
      - in: "query"
        name: "orgId"
        required: false
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CapacityTrendResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-summary-controller"
...
  /admin/summary:
    get:
      operationId: "getAdminSummary"
      parameters:
      - in: "query"
        name: "orgId"
        required: false
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrgDashboardSummaryResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-summary-controller"
  /admin/system-summary:
    get:
      operationId: "getSystemSummary"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SystemDashboardSummaryResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-summary-controller"
  /admin/tasks:
    get:
      operationId: "listAdminTasks"
      parameters:
      - in: "query"
        name: "status"
        required: false
        schema:
          items:
            $ref: "#/components/schemas/ProvisioningTaskStatus"
          type: "array"
      - in: "query"
        name: "kind"
        required: false
        schema:
          $ref: "#/components/schemas/ProvisioningTaskKind"
      - in: "query"
        name: "vmId"
        required: false
        schema:
          format: "uuid"
          type: "string"
      - in: "query"
        name: "page"
        required: false
        schema:
          default: 0
          format: "int32"
          minimum: 0
          type: "integer"
      - in: "query"
        name: "size"
        required: false
        schema:
          default: 20
          format: "int32"
          maximum: 100
          minimum: 1
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseAdminTaskResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-task-controller"
  /admin/tasks/{taskId}/retry:
    post:
      operationId: "retryAdminTask"
      parameters:
      - in: "path"
        name: "taskId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: "Accepted"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-task-controller"
...
  /admin/users:
    get:
      operationId: "listUsers"
      parameters:
      - in: "query"
        name: "q"
        required: false
        schema:
          maxLength: 100
          minLength: 1
          type: "string"
      - in: "query"
        name: "status"
        required: false
        schema:
          $ref: "#/components/schemas/UserStatus"
      - in: "query"
        name: "role"
        required: false
        schema:
          $ref: "#/components/schemas/UserRole"
      - in: "query"
        name: "orgId"
        required: false
        schema:
          format: "uuid"
          type: "string"
      - in: "query"
        name: "sort"
        required: false
        schema:
          enum:
          - "name"
          - "-name"
          - "email"
          - "-email"
          - "createdAt"
          - "-createdAt"
          type: "string"
      - in: "query"
        name: "page"
        required: false
        schema:
          default: 0
          format: "int32"
          minimum: 0
          type: "integer"
      - in: "query"
        name: "size"
        required: false
        schema:
          default: 20
          format: "int32"
          maximum: 100
          minimum: 1
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseUserAdminViewResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-user-controller"
  /admin/users/{userId}:
    get:
      operationId: "getUser"
      parameters:
      - in: "path"
        name: "userId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserAdminDetailResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-user-controller"
    patch:
      operationId: "updateUser"
      parameters:
      - in: "path"
        name: "userId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserAdminRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserSummaryResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-controller"
  /admin/users/{userId}/disable:
    post:
      operationId: "disableUser"
      parameters:
      - in: "path"
        name: "userId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisableUserRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserAdminDetailResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-user-controller"
  /admin/users/{userId}/enable:
    post:
      operationId: "enableUser"
      parameters:
      - in: "path"
        name: "userId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserAdminDetailResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-user-controller"
  /admin/users/{userId}/mfa-reset:
    post:
      operationId: "resetUserMfa"
      parameters:
      - in: "path"
        name: "userId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-user-controller"
...
  /me/mfa/disable:
    post:
      operationId: "disable"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisableMfaRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "mfa-controller"
...
security:
- bearerAuth: []
servers:
- url: "/api/v1"
...